buggig version av mapsync

+++ TDDD36Server/src/tddd36/server/Sender.java
@@ -105,6 +105,7 @@ public class Sender implements Runnable {
 			// Logga ut denna användaren om 
 			LoginManager.logoutUser(((Message) messages.get(0)).getDestUser());
 			for(Object m : messages){
+				if(m instanceof RequestMessage) continue;
 				Database.storeIntoBuffer((Message)m);
 				Database.deleteFromTextMessages((TextMessage)m);
 			}
